+++ src/main/kotlin/racinggame/car/domain/Engine.kt
@@ -1,4 +1,4 @@
-package racinggame.car
+package racinggame.car.domain
 
 interface Engine {
     fun enoughPower(): Boolean
